% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fixedeffect.R
\name{est_fixedeffect}
\alias{est_fixedeffect}
\title{est_fixedeffect}
\usage{
est_fixedeffect(vec_logitc, vec_SElogitc)
}
\arguments{
\item{vec_logitc}{Array of logit transformed c-statistics (AUC) of predictive model studies. Calculated in order using the logit function.}

\item{vec_SElogitc}{Array of standard errors associated with the logit transformed c-statistics (AUC) of each predictive model study. Calculated using the est_SElogitc function.}
}
\value{
Pooled estimate of the c-statistic (fixed effects), value from 0 to 1.
}
\description{
Calculates estimate (fixed effects) of the c-statistic for multiple pooled prediction model studies
}
