% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_SEc.R
\name{est_SEc}
\alias{est_SEc}
\title{est_SEc}
\usage{
est_SEc(c, num_events, num_non_events)
}
\arguments{
\item{c}{C-statistic (AUC) of a predictive model, value between 0 and 1}

\item{num_events}{Number of events in the dataset used to generate the c-statistic}

\item{num_non_events}{Number of non-events in the dataset used to generate the c-statistic (number of cases - number of events)}
}
\value{
Standard error of the c-statistic
}
\description{
Estimate standard error of the c-statistic
}
