% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_varlogitc.R
\name{est_varlogitc}
\alias{est_varlogitc}
\title{est_varlogitc}
\usage{
est_varlogitc(c, num_events, num_non_events)
}
\arguments{
\item{c}{C-statistic (AUC) of a predictive model, value between 0 and 1}

\item{num_events}{Number of events in the dataset used to generate the c-statistic}

\item{num_non_events}{Number of non-events in the dataset used to generate the c-statistic (number of cases - number of events)}
}
\value{
Variance of the logit of the c-statistic
}
\description{
Estimate variance of the logit of the c-statistic based on the sample size and number of events.
}
