% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\title{invlogit}
\usage{
invlogit(logitx)
}
\arguments{
\item{x}{A continuous value}
}
\value{
The inverse logit, a value between 0 and 1
}
\description{
Inverse of the logit function
}
