% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_varlogitc.R
\name{est_varlogitc}
\alias{est_varlogitc}
\title{est_varlogitc}
\usage{
est_varlogitc(c, num_events, num_non_events)
}
\arguments{
\item{x}{A value between 0 and 1}
}
\value{
The logit of x
}
\description{
Testing 123
}
