% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI_high.R
\name{get_CI_high}
\alias{get_CI_high}
\title{get_CI_high}
\usage{
get_CI_high(logitc, varlogitc, n, alpha = 0.05)
}
\arguments{
\item{x}{A value between 0 and 1}
}
\value{
The logit of x
}
\description{
Testing 123
}
