% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI_low.R
\name{get_CI_low}
\alias{get_CI_low}
\title{get_CI_low}
\usage{
get_CI_low(logitc, varlogitc, n, alpha = 0.05)
}
\arguments{
\item{logitc}{logit of the c-statistic (AUC) of a predictive model, calculated using the logit function}

\item{varlogitc}{variance of the logit of the c-statistic (AUC) of a predictive model, estimated using est_varlogitc}

\item{n}{Sample size in the dataset used to generate the c-statistic}

\item{alpha}{Desired significance level}
}
\value{
Lower bound of confidence interval of the c-statistic
}
\description{
Estimate lower bound of confidence interval of the the c-statistic.
}
