% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI_vec.R
\name{get_CI_vec}
\alias{get_CI_vec}
\title{get_CI_vec}
\usage{
get_CI_vec(logitc, SElogitc, n, alpha = 0.05)
}
\arguments{
\item{logitc}{logit of the c-statistic (AUC) of a predictive model, calculated using the logit function}

\item{SElogitc}{standard error of the logit of the c-statistic (AUC) of a predictive model, estimated using est_SElogitc}

\item{n}{Sample size in the dataset used to generate the c-statistic}

\item{alpha}{Desired significance level}
}
\value{
Lower bound of confidence interval of the c-statistic
}
\description{
Estimate confidence intervals of the the c-statistic for multiple studies at once. Inputs are arrays representing corresponding values for each study.
}
